+++ components/select/style/index.ts
@@ -4,3 +4,4 @@ import '../../select-trigger/style/index.less';
 import '../../scrollbar/style/index.less';
 import '../../popper/style/index.less';
 import '../../ellipsis/style/index.less';
+import '../../virtual-list/style/index.less';