+++ components/table/interface.ts
@@ -5,63 +5,15 @@ import useTableEvent from './useTableEvent';
 import useTableSelect from './useTableSelect';
 import useTableExpand from './useTableExpand';
 import useTableStyle from './useTableStyle';
+import type { TableProps } from './table';
 
 export type RowType = Record<string, any>;
-
-export type RowKey = string | ((row: RowType) => string);
+export type RowKey = string | ((row: RowType) => string | number);
 
 export interface ActionType {
     label: string | number;
     func: (row: object) => void;
 }
-
-// export interface ColumnProps {
-//     label: string;
-//     prop: string;
-//     type: typeof COL_TYPE[number];
-//     align: typeof ALIGN[number];
-//     width: number;
-//     minWidth: number;
-//     colClassName:
-//         | string
-//         | (({
-//               row,
-//               column,
-//               rowIndex,
-//               columnIndex,
-//               cellValue,
-//           }: {
-//               row: RowType;
-//               column: ColumnInst;
-//               rowIndex: number;
-//               columnIndex: number;
-//               cellValue: any;
-//           }) => string);
-//     colStyle:
-//         | object
-//         | (({
-//               row,
-//               column,
-//               rowIndex,
-//               columnIndex,
-//               cellValue,
-//           }: {
-//               row: RowType;
-//               column: ColumnInst;
-//               rowIndex: number;
-//               columnIndex: number;
-//               cellValue: any;
-//           }) => object);
-//     fixed: 'left' | 'right' | true | false;
-//     formatter: (data: CellProps) => any;
-//     resizable: boolean;
-//     sortable: boolean;
-//     sortMethod: () => void;
-//     selectable: () => void;
-//     action: ActionType | ActionType[];
-//     ellipsis: boolean;
-//     visible: boolean;
-// }
 
 export interface TableInst
     extends ReturnType<typeof useTableColumn>,
@@ -69,6 +21,7 @@ export interface TableInst
         ReturnType<typeof useTableSelect>,
         ReturnType<typeof useTableExpand>,
         ReturnType<typeof useTableStyle> {
+    rootProps: TableProps;
     getRowKey: ({ row }: { row: RowType }) => string | RowType;
     getCellValue: typeof getCellValue;
     tableId: string;