+++ components/version/index.ts
@@ -1,3 +1,3 @@
-const version = '0.3.1';
+const version = '0.3.3';
 
 export default version;